input

+++ src/main/java/diplom/model/User.java
@@ -13,7 +13,7 @@ public class User implements UserDetails {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private int id;
 
     private String username;
     private String LNAME;
@@ -39,9 +39,9 @@ public class User implements UserDetails {
         this.roles = roles;
     }
 
-    public Long getId() { return id; }
+    public int getId() { return id; }
 
-    public void setId(Long id) { this.id = id; }
+    public void setId(int id) { this.id = id; }
 
     public void setUsername(String username) { this.username = username; }
